Move `mustBeDefined` into `util` in graph
