--- libs/@blockprotocol/graph/src/shared/stdlib/must-be-defined.ts
@@ -1,7 +1,0 @@
-export const mustBeDefined = <T>(x: T | undefined, message?: string): T => {
-  if (x === undefined) {
-    throw new Error(`invariant was broken: ${message ?? ""}`);
-  }
-
-  return x;
-};
