+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/edge/link-entity.ts
@@ -13,13 +13,12 @@ import {
   Subgraph,
 } from "../../../types/subgraph.js";
 import { TimeInterval } from "../../../types/temporal-versioning.js";
-import { typedEntries } from "../../../util.js";
+import { mustBeDefined, typedEntries } from "../../../util.js";
 import {
   intervalForTimestamp,
   intervalIntersectionWithInterval,
   intervalIsStrictlyAfterInterval,
 } from "../../interval.js";
-import { mustBeDefined } from "../../must-be-defined.js";
 import {
   getEntityRevision,
   getEntityRevisionsByEntityId,
